Enable EDNS0 in queries
This allows receipt of responses >512 bytes in length.

If a recursive doesn't support EDNS, the stub will retry without
EDNS.  This logic was already in place for gethostbyname, but
making it work for getaddrinfo required porting res_nquery's retry
logic to res_queryN, which was forked from res_nquery before the
retry logic was added.  This change has also landed in NetBSD
upstream (http://gnats.netbsd.org/52578) so this change reduces
divergence from upstream.

Most of the code in this change is for caching of queries that contain
an additional section.

Bug: 15132200
Test: Added integration tests for fallback to the netd suite.
Change-Id: Ic64bed0754e1d529dc0c0ab6a5e2f1ea201ff0d5