Update patch set 3
Patch Set 3:

> Patch Set 3:
> 
> > Patch Set 3:
> > 
> > > Patch Set 3:
> > > 
> > > > Patch Set 3:
> > > > 
> > > > Adding __asm__("mmap64") to a declaration of the static inline mmap64 could be one way to fix the Clang C++ problem.
> > > 
> > > This was the first thing I'd tried, but maybe the error I got was actually from GCC?
> > 
> > Looks like __asm__ isn't allowed on function definitions, only decls.
> 
> Adding it to a declaration before the definition worked for me:
> 
> static inline void bar(void) __asm__("bar");
> static inline void bar(void) {
>     puts("bar");
> }

That's weird.

> You were able to compile an mmap call Clang C++ though without this __asm__ change, though?

Yes. I changed the __RENAME in sys/mman.h to use the mangled name. The weird forward decl of the inline function seems better than this though, so I'll change it.

> Does ndk-build's APP_CFLAGS apply to C++?

Yep. There's a CFLAGS, CPPFLAGS, a CXXFLAGS, and a CONLYFLAGS. CFLAGS are applied to C and C++, CONLYFLAGS are specifically C.

Patch-set: 3