Implement pthread_attr_getinheritsched/pthread_attr_setinheritsched.
Also remove a comment about setting the scheduler attributes in main_thread
because we'd have to actually keep them up to date, and it's not clear that
doing so would be worth the trouble.

Also make async_safe_format_log preserve errno so we don't have to be
so careful around it.

Bug: http://b/67471710
Test: ran tests
Change-Id: Idd026c4ce78a536656adcb57aa2e7b2c616eeddf