RTLD_NEXT does not skip global group
Previously, RTLD_NEXT starts searching from lib next to the calling
library. As a result, libs loaded before the calling library were
skipped even if they are in the global group (RTLD_GLOBAL or
LD_PRELOAD'ed).

This behavior is not in sync with the standard[1] and as a result
dlfcn.rtld_next_from_library fails when a lib having close() is in
LD_PRELOAD.

Fix the behavior by skipping a lib only when it is in front of the
calling lib and not in the global group.

Bug: 67978141
Test: dlfcn.rtld_next_with_df_1_global

Change-Id: I26901d3e49102106767c508d77dabd2c5ba05ae0