Update patch set 2
Patch Set 2:

With the LLNDK, as long as you're linking to libc, none of this really matters (and nearly everyone at that level is linking to libc if they're also linking one of the others). It's really just the ABI accounting that Jayant is doing.

Patch-set: 2
Reviewer: Gerrit User 1074628 <1074628@85c56323-6fa9-3386-8a01-6480fb634889>