Update patch set 2
Patch Set 2:

did you test it against glibc linker? bionic-unit-tests-glibc?

I do not think your interpretation is correct. The point of this sentence below is not that RTLD_GLOBAL should be looked up even if it comes before. the point is that RTLD_NEXT should NOT look into other load groups unless they have libraries loaded after 'this' one and have RTLD_GLOBAL flag.

"The next object is either one of global scope (because it was introduced as part of the original process image or because it was added with a dlopen() operation including the RTLD_GLOBAL flag), or is an object that was included in the same dlopen() operation that loaded this one."

Patch-set: 2
Reviewer: Gerrit User 1041673 <1041673@85c56323-6fa9-3386-8a01-6480fb634889>