bionic: add vdso clock_getres and time
clock_getres() should not be a hot call, nevertheless it is
~7 times faster for supported clock ids if it uses
__vdso_clock_getres if available.  There is a 15% performance
penalty for unsupported clock ids via __vdso_clock_getres, but
none if the kernel does not supply it.

time() can be a hot call, and it currently uses __vdso_gettimeofday,
which is already pretty fast (~2.5 times faster than the syscall),
but with a __vdso_time call it is ~5 times even faster, in part
because __vdso_time does not require interlocking with updates,
and the read for just the seconds is atomic.

Following is measured on a hikey960 locked at full speed, on the
little cores, for vdso32.

               Before    After
clock_getres   165ns     23ns (191ns for unsupported clock ids)
time           92ns      19ns

Test: liblog-benchmarks BM_clock_getres_overhead \
                        BM_clock_getres_syscall_overhead \
                        BM_time_overhead
Bug: 63737556
Change-Id: I80c0a5106625d76720287f715fcf145d2aad1705