Update patch set 5
Patch Set 5:

It appears I need to provide reasons for the exceptional values. They are not tests, but baselines. In 'bionic: add vdso time' results (which used to be in the commit here):

syscall-unsupported=175-100=75ns to process+calculate CLOCK_MONOTONIC inside the kernel

TAI-syscall=210-175=35ns approx time to process vdso overhead

gettime-vdso-overhead=48-35=13ns to _acquire_ and _calculate_ CLOCK_MONOTONIC using vdso

Adding unsupported to the list just gave me the ability to measure the posix_timers registration and selection overhead that the system calls use to see it was 75ns-13ns=62ns to scope through. 62ns is such a waste of time rather than going directly to the registers like we do in the vdso! Maybe we can fix the timer handling in the kernel, because this overhead may be incurred now for every printk. Imagine, I can actually get the time faster in user space, than someone inside the kernel can get it(!)

And so on, I used the figures in development.

Focused on reducing the CLOCK_MONOTONIC calculation time as you see above and determined I could lower it on vdso32 by judiciously using u32 and by moving the calculations outside the atomic loop, just pick up the key values and do the calculations once on the way out. This had to have _zero_ performance impact in the general case, but made sense as being more efficient in the specific case when it had to loop. Further still by dropping some fixed values (fixed mask on vdso32 saved time, whereas on arm we have to program in a variable mask adding one more variable to pick up). All the optimizations managed to be generalized on common arch config options.

Very key to development and optimization, and I expect that the combinations of these values and how they inflect will tell me _where_ a timing regression got introduced.

Yes, it may be overboard post-development, so I am willing to accept that some of these tests could be available, but turned off in the full.xml case?

Patch-set: 5
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>