Update patch set 1
Patch Set 1: Code-Review+2

Added Colin and Dan to check that there isn't a better way to parse go in python (but I am pretty sure that this is the best we are going to do). It's certainly better than the previous version that was getting lucky.

One possible alternative is to actually invoke the build system to get the information and/or have the build system actually produce the artifact that you need (i.e. run clang to just preprocess the kernel header). Then you could just read the resulting file from the output directory (and it would always be built during checkbuilds, so less worry about it bit-rotting).

Patch-set: 1
Reviewer: Gerrit User 1010056 <1010056@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2