Update patch set 1
Patch Set 1:

This is depending entirely on implementation details and is not an appropriate way of getting data out of Soong.  I'd suggest making the script take the path to clang as the first argument, and then making a rule in a makefile that uses the build system's knowledge of clang to run the script.

Patch-set: 1
Reviewer: Gerrit User 1002751 <1002751@85c56323-6fa9-3386-8a01-6480fb634889>