Fix bug with double unload on unsuccessful dlopen
In the case of unsuccessful dlopen the failure guard is triggered
for two namespaces which leads to double unload.

In addition, fix a bug that, when two libs libFoo and libBar are both
loaded in a linked namespace and libFoo -> libBar, libBar's reference
count is not incremented. Ref count is now incremented when a lib being
loaded is needed by other libs that are to be loaded in the current run.

The second test demonstrates the problem with reference counting
introduced by 02586a2a34e6acfccf359b94db840f422b6c0231.

Also update soinfo_free to abort in case when linker tries to free same
soinfo for the second time - this makes linker behavior less undefined.

Test: bionic-unit-tests
Bug: http://b/69787209

Change-Id: Ic5324c69063132bbeff176e10ea7639b7c552e0f