Update patch set 8
Patch Set 8:

> Patch Set 8:
> 
> > How do I add a benchmark test that does not run in the 'full' case
>  > because it provides refined details that few care about, but does
>  > run under a selected suite like vdso.xml now?
> 
> you don't. we don't think that's appropriate, as discussed already :-)
> 
> we think benchmarks are for performance cases, not meant to cover every possible input including errors *unless* that's been an actual "this caused a problem for real code" problem. (something like mktime, say, could legitimately fail a lot in a real app if it's dealing with user input.)

Will need to keep a private branch of the bionic tests so that we can measure the various primary and differential code paths, error cases acting as a baseline, in order track where the time and overhead is spent in the kernel. Originally I had private tests in liblog's benchmark, and moved them into bionic to standardize and make it easier to see them on a dashboard. The price I pay for having a kernel benchmark in libc.

Would it be acceptable if I added these interesting kernel cases behind an #ifdef FULL_VDSO_BENCHMARKS so that future changes do not result in continuing maintenance (for me), of course leaving the globally interesting set as the default? We will need to add support for listing a test in the vdso.xml that has been ifdef'd out though...

Patch-set: 8
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0