Update patch set 12
Patch Set 12:

Can we get a +2 on the pair of benchmark changes this depends on so we can merge this set of three (benchmark clock_gettime, benchmark clock_getres and this patch)?

b/69959622 (marlin kernel with faulty clock_getres) has not moved yet, but expect it to shortly. If not, this will light a fire. taimen is ready.

Patch-set: 12
Reviewer: Gerrit User 1032276 <1032276@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=0
Label: Presubmit-Ready=0