Fix crash after running system properties tests/benchmarks
System properties tests and benchmarks don't handle restoring the
original property area currently.  This is particularly problematic
when caching a prop_info pointer as the pointer gets invalidated,
which breaks a fundamental guaranteed of system properties.

This change puts a hack into __system_properties_set_filename() that
backs up the system property areas when changing the filename away
from PROP_FILENAME (the system property location) and restores it when
changing the filename back to PROP_FILENAME.

Bug: 62197783
Test: new unit test for this specific issue, bionic benchmarks
Change-Id: I998a1c02f5a152caad231ef6fbe645770f9f5a6b