Fix dlclose for libraries with thread_local dtors
Mark library NODELETE once first pthread_atexit handler is registered.

According to the test runs it seems that glibc loader agrees on the fact
that it is too dangerous to unload a library on pthread_exit(). It only
does it on dlclose if there are no active thread local dtors left
associated with the dso.

(At this point we just have tests to demonstrate how it works)

Test: bionic-unit-tests --gtest_filter=dl*
Test: bionic-unit-tests-glibc --gtest_filter=dl*
Bug: https://github.com/android-ndk/ndk/issues/360
Change-Id: I724ef89fc899788f95c47e6372c38b3313f18fed