linker: Add allow_all_shared_libs to ld.config.txt format
This commit adds allow_all_shared_libs to ld.config.txt file format so
that it will be possible to link two linker namespaces without the
soname filter.

The motivation is to establish one-way shared library isolation.  For
example, assume that there are two linker namespaces a and b.  We would
like to limit the shared libraries that can be used by linker namespace
b.  In the meanwhile, we would like to allow linker namespace a to use
all shared libraries that are loaded in linker namespace b.

             shared_libs = ...
    ns_a <----------------------------- ns_b
       \_______________________________/^
          allow_all_shared_libs = true

In this CL, allow_all_shared_libs is only available to ld.config.txt and
is not exposed in any public APIs.

Bug: 69824336

Test: adb shell /data/nativetest64/linker-unit-tests/linker-unit-tests64

Test: Update /system/etc/ld.config*.txt and check whether the vndk
linker namespace of the vendor process can access the shared libs from
the default linker namespace.

Change-Id: I2879f0c5f5af60c7e56f8f743ebd2872e552286b