linker: Allow link namespaces without name filters
This commit allows users to create a link without soname filters between
two linker namespaces.

The motivation is to establish one-way shared library isolation.  For
example, assume that there are two linker namespaces a and b.  We would
like to limit the shared libraries that can be used by linker namespace
b.  In the meanwhile, we would like to allow linker namespace a to use
all shared libraries that are loaded in linker namespace b.

          shared_libs = liba.so:libb.so
    ns_a <----------------------------- ns_b
       \_______________________________/^
          shared_libs = /*all-libs*/

Bug: 69824336

Test: adb shell /data/nativetest64/linker-unit-tests/linker-unit-tests64

Test: adb shell /data/nativetest64/bionic-unit-tests/bionic-unit-tests

Test: Update /system/etc/ld.config*.txt and check whether the vndk
linker namespace of the vendor process can access the shared libs from
the default linker namespace.

Change-Id: I2879f0c5f5af60c7e56f8f743ebd2872e552286b