Update patch set 2
Patch Set 2:

(1 comment)

What is the point where the deadlock occurs?  The only lock in the __system_property_* functions is in ContextNode::Open().  Is this lock causing the deadlock?  What is it contending with?

How does moving the initialization outside of the block help?

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>