Update patch set 4
Patch Set 4:

Proposal: can we allocate the first-level lookup table in BSS so that we don't have to load the top-level pointer to manipulate the lock? (We can find the location of BSS by PC arithmetic.) We can align the table so that it gets its own BSS page, which we'll fault only when a program actually tries to use 32-bit PI mutexes. (To avoid introducing BSS fragmentation that increases overall memory use, we can use a linker script to put this page last.)

This way, manipulating PI locks becomes cheaper, but it's still pay-for-what-you-use, except for the address space reservation, which is nearly free.

Patch-set: 4