Update patch set 4
Patch Set 4:

> Patch Set 4:
> 
> > Although it is possible, I think the performance gain doesn't outweigh the complexity added to the implementation. On pixel 2, the benchmark result of BM_pthread_mutex_lock_PI goes from 89 ns to 87 ns after removing loading the top-level pointer, which means only 2% gain.
> 
> Thanks for checking. Wouldn't the implementation actually be simpler in the bss case though? You'd avoid a call to malloc and just use a pre-built array.

In your plan (align the table so that it gets its own BSS page, which we'll fault only when a program actually tries to use 32-bit PI mutexes. To avoid introducing BSS fragmentation that increases overall memory use, we can use a linker script to put this page last.), adding a linker script to put this page last is the most complex part. I doubt if adding linker script is supported by Android build system, and it's hard to maintain.

Patch-set: 4