Avoid hackery in mksh.
I had trouble with the last mksh upgrade because I dropped a local patch
there. If we add an #if !defined in bionic, we can replace the mksh code
change with build configuration instead.

This patch makes me realize that we already have a bug here:
execvp/execvpe will use the default path that makes sense for binaries
on /system even if called from /vendor. We'll need to do something
similar to what we already do for _PATH_BSHELL.

(The fact that the current _PATH_BSHELL workaround is done for
upstream-netbsd code only also means that execve/execvpe's code for
exec'ing a script is broken for /vendor, and the rewrite of system(3)
that I'm currently working on would have the same problem.)

Bug: N/A
Test: treehugger
Change-Id: Ib0899bcfa81bcd1324d6d0900f275584bed7db7d