Avoid hackery in mksh.
I had trouble with the last mksh upgrade because I dropped a local patch
there. We already had a workaround in bionic for _PATH_BSHELL in system(3)
and popen(3), and should just generalize and expose them for all platform
code. In particular, this patch adds similar code for _PATH_DEFPATH.

(The fact that the old _PATH_BSHELL workaround was done for
upstream-netbsd code only also meant that execve/execvpe's code for
exec'ing a script was broken for /vendor, and the rewrite of system(3)
that I'm currently working on would have the same problem.)

Bug: http://b/67975799
Test: treehugger
Change-Id: Ib0899bcfa81bcd1324d6d0900f275584bed7db7d