Update patch set 9
Patch Set 9: Presubmit-Ready+1

(1 comment)

I cleaned up the interface a bit.

Regarding what to do with /system/etc/{passwd,group}, this code is generic enough that we can trivially support adding parsing of that file, but that can be done as a future task in a separate CL.  This CL will directly address the concern of reading from /vendor, especially due to the need for that given treble.  

/odm can also be addressed in the future if/when that need arrises.

I think this is good to go.  I'm interested in hearing ack's from other reviewers if this will fulfill their needs.

Patch-set: 9
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1