Update patch set 2
Patch Set 2:

> Patch Set 2: Code-Review+1
> 
> This is fine with me.  Erik, please let me know if you think this change obsoletes https://android-review.googlesource.com/c/platform/system/netd/+/602913
> 
> (We can certainly have both, but it's not clear why we would.)

My preference would be for passing explicit signals into the resolve innards.

TLS-or-not is only 1 bit of information.  For now that means we can turn that into: EDNS=true and later add DNSSEC=true, and so on.  But we should also have the ability to do each of these special things, and any future special things, on their own.

So I think in netd, when we observe TLS we turn on all the magic we want.  We might want to turn on things in other circumstances, eventually.

Patch-set: 2
Reviewer: Gerrit User 1000451 <1000451@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0