Update patch set 2
Patch Set 2: Code-Review+1

Flag LGTM.

However, I don't know if this will result in this flag being visible to NDK apps. Before we expose it to such apps, we should ensure that we're all agreed that that is the solution we want.

Another option would be to add a variant of android_getaddrinfofornetwork that takes a uint32_t flags.

Patch-set: 2
Reviewer: Gerrit User 1000835 <1000835@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+1