Update patch set 1
Patch Set 1:

We were exporting the symbol for 14 and 15 in libc.map.txt until very recently:

https://android-review.googlesource.com/c/platform/bionic/+/633629/

I think that CL used 16 because it was the minimum version. It's only the minimum for NDK r18, though, and we updated the platform headers/stubs for NDK r17.

Maybe we should reduce the API from 16 to 14 in libc.map.txt?

Patch-set: 1
Reviewer: Gerrit User 1229339 <1229339@85c56323-6fa9-3386-8a01-6480fb634889>