Update patch set 1
Patch Set 1:

have you thought about how we'll keep this up to date? i see that upstream is actually active (which is a good thing!) so i wonder whether we should try to set up external/ARM-software-optimized-routines, build a .a file (and a gtest) there, and link that into libm?

i'm assuming you're not signing up to maintain a copy of the upstream code in bionic for life? :-)

Patch-set: 1
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>