Update patch set 5
Patch Set 5:

> Patch Set 5:
> 
> For AArch64, at least, it means now both memmove and memcpy shares the same entrypoint and each call to memcpy add an extra 4 instructions and a taken branch to every call. From previous comments it seems the performance drop was acceptable, without further explanation why.
> 
> Since neither the standard or other libc implementation guarantee such behavior,

you're right that glibc doesn't -- and has actually moved in the opposite direction, much to Linus' annoyance -- but other important libcs do.

> seems sloppy imho to push such behavior. Compiler will still conservatively auto-generate memcpy calls only for non-overlap region (unless someone adds flags to change its behavior, but it seems not a quite profitable flag since memmove is exactly the way for such) 

we're not worried about the _compiler_ making incorrect assumptions. it's humans we're trying to defend against :-)

you'll see this a lot in bionic: from extra bounds checks that the standard doesn't mandate to things like explicit null pointer checks that give clear error messages when they abort rather than just SIGSEGV.

correctness and security are more important than tweaking spec scores.

> and the implementation of memcpy/memmove could still be optimized (in both code size and complexity) by using a single merged function with 2 entry points (since usually memove internally branches to memcpy for non-overlap memory regions).

yes please. i'll +2 that patch :-)

> So the question is why this behavior is preferred to Bionic?

the reason for the _behavior_ change is obvious from the philosophical point above.

the reason for the effort put into the _implementation_ is that (a) that kind of thing is largely left to SoC vendors anyway and (b) this is also a small part of a much larger project we're working on for Q, where we'll switch to ifuncs and runtime selection of optimized routines rather than the build-time selection you see today. since "make memcpy safe" has been on our to-do list for years and slightly reduces the amount of work we need to do with the ifuncs, we did this first.

Patch-set: 5
Reviewer: Gerrit User 1003224 <1003224@85c56323-6fa9-3386-8a01-6480fb634889>