Enable membarrier syscall for system_server
The membarrier syscall could be used to synchronize the rcu inside the
kernel. It is useful when userspace and kernel read/write the eBPF maps
at the same time. The userspace uses it to garantee the change to eBPF
maps is updated on all CPUs and then system_server can do map
modifications such as cleaning and scraping.

Bug: 79171384
Test: device boot into UI without seccomp crash

Change-Id: Icea23616ed20b63d98666e66c3ad16dbdaccb8e3