Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > > Patch Set 1:
> > > 
> > > > Patch Set 1:
> > > > 
> > > > > Patch Set 1:
> > > > > 
> > > > > > Patch Set 1:
> > > > > > 
> > > > > > Thanks. I agree that an explicit function of this sort in BPF is a good idea. It's explicit about what it guarantees, and allows the implementation to change in the future. Hopefully upstream will agree.
> > > > > 
> > > > > Thanks! That sounds much better to me.
> > > > 
> > > > Me too. Unfortunately, the proposal is getting a little resistance upstream. Anyone want to chime in? 
> > > > 
> > > > https://marc.info/?l=linux-netdev&m=153093379520439&w=2
> > > 
> > > I kind of agree with Alexei on that one. Do you really feel MEMBARRIER_CMD_GLOBAL can really be implemented in any other way other than synchronize_sched() ?
> > 
> > Hans mentioned some architectures (some nVidia thing?) having fast hardware support for the equivalent operation.
> > 
> > > One thing against membarrier could be that it does this:
> > > 
> > >                 /* MEMBARRIER_CMD_GLOBAL is not compatible with nohz_full. */
> > >                 if (tick_nohz_full_enabled())
> > >                         return -EINVAL; 
> > > 
> > > So that means membarrier method may not work on a nohz-full system?
> > 
> > Good point. And since BPF programs execute _only_ in the kernel, BPF_SYNCHRONIZE would have the intended effect even on NOHZ_FULL programs.
> > 
> > Besides: suppose that we did spec MEMBARRIER_CMD_GLOBAL to mean synchronize_sched. How would you talk about such a guarantee in the man page? The spec doesn't and shouldn't talk about things userspace doesn't care about, like RCU grace periods. Would you just say "Oh, and by the way, this command also has the effect of blah-and-blah on BPF programs."? It'd be a very strange interface.
> > 
> > In any case, what's the downside? The implementation is trivial and the semantics clear.
> 
> I think the downside is just adding a new interface to BPF. Other than that I don't see any.
> 
> I think the points you/I brought up can be brought up on that thread and see if Alexei is convinced.

Right. I think I made my case though, so I'd appreciate if someone else commented too.

> Plus I think the point about membarrier() needing a matching barrier is also a good one. May I ask you though, just for my knowledge - what happens if a memory barrier isn't matched with a counterpart? Is there a potential "issue" that can arise? I know the canonical law is that it should be matched if the architecture needs it to be, but I was always curious about the real reason about what can go wrong.

Hans was concerned about the unpaired membarrier because, as a general code review heuristic, synchronization on only on the producer side of a producer-consumer system indicatives a design flaw. Totally free read and write reordering done by hardware and software can produce write-sequence observations that appear to violate logic and causality. We want to prohibit such reorderings, and we do that by inserting barriers across which memory operations can't move. It just so happens that on many popular systems, since both reads and writes can both be reordered, we need barriers in both the memory-writing and the memory-reading code in order in order to limit the observable write sequences to the ones we want. That's why it's usually a bug when you see a barrier in just the reader or just the writer.

There's no "issue" that comes from the use of an unpaired barrier by itself. (It's not as if we have barrier tokens that can leak or something.) In certain special applications, such a thing might even be correct. In our case, we don't care about the userspace memory barrier: we just want to take advantage of the side effect of the implementation, and this side effect (the RCU subsystem) already has properly constrained concurrency.

Patch-set: 1
Reviewer: Gerrit User 1130394 <1130394@85c56323-6fa9-3386-8a01-6480fb634889>