Update patch set 1
Patch Set 1: Presubmit-Ready+1

I was wondering why we didn't always do it this way? It seemed obvious...

This change does require the linker of the loader to emit a PT_PHDR segment, and bfd only emits PT_PHDR for binaries that have an .interp section. The loader is a shared object with no .interp/PT_INTERP, so with this patch, if it's linked with bfd, then it can't be run by itself. (It would segfault.)

gold and lld, on the other hand, emit PT_PHDR unconditionally.

Patch-set: 1
Reviewer: Gerrit User 1229339 <1229339@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1