Update patch set 2
Patch Set 2:

(1 comment)

Would it be better to implement this ourselves using __builtin_mul_overflow instead of the openbsd code?

In fact, putting that code directly in malloc_common.cpp might be better since all of the malloc functions would be in one place.

Although, it's possible there would be some linking if that file is not used with the openbsd code.

Patch-set: 2
CC: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>